backend/requirements.txt
numpy
pandas
matplotlib
scikit-learn
joblib
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi
uvicorn
pydantic
